requirements.txt
# Telegram bot
python-telegram-bot>=22.0

# Gemini 3 Flash (google-genai)
google-genai>=1.51.0

# HTTP for OpenRouter fallback
httpx>=0.27.0

# Config
pydantic-settings>=2.0.0

# Logging
loguru>=0.7.0

# Optional: resize reference photos for image API (get_reference_photo_bytes)
Pillow>=10.0.0
